refactor: deduplicate pipeline functions and add FollowupRampConfig
- Extract shared pipeline functions (subsample, derive_features,
  softcap_expected_acv, assign_acquisition_wave, rename_and_select,
  inject_missingness_v6) into leadforge/pipelines/common.py
- Extract shared ML pipeline (build_baseline_pipeline, build_preprocessor,
  fit_evaluate, get_feature_cols, sanitize_categoricals) into
  leadforge/pipelines/ml.py
- Define CAT_FEATURES, NUM_FEATURES, BINARY_FEATURES once in common.py;
  validators and eval scripts import from there
- Introduce FollowupRampConfig frozen dataclass in mechanisms/counts.py;
  LatentDecayIntensity accepts followup param (legacy params still work)
- Fix subsample() to raise ValueError when insufficient negatives
  instead of silently returning a short DataFrame

All 866 tests pass, lint clean.
