+++ leadforge/pipelines/build_v7.py
@@ -15,12 +15,24 @@ Key v7 changes over v6:
 
 from __future__ import annotations
 
-import warnings
-
-import numpy as np
 import pandas as pd
 
-from leadforge.core.rng import RNGRoot
+from leadforge.pipelines.common import (
+    ACV_CAP,
+    ACV_FLOOR,
+    SUBSAMPLE_N,
+    TARGET_RATE,
+    assign_acquisition_wave,
+    derive_features,
+    softcap_expected_acv,
+    subsample,
+)
+from leadforge.pipelines.common import (
+    inject_missingness_v6 as inject_missingness,
+)
+from leadforge.pipelines.common import (
+    rename_and_select as _rename_and_select_generic,
+)
 
 __all__ = [
     "ACV_CAP",
@@ -49,12 +61,6 @@ __all__ = [
 SEED = 42
 N_LEADS = 5000
 SNAPSHOT_DAY = 20
-SUBSAMPLE_N = 1000
-TARGET_RATE = 0.30
-
-# Narrative-consistent ACV bounds (from narrative.yaml: $18k-$120k).
-ACV_FLOOR = 18_000.0
-ACV_CAP = 120_000.0
 
 INSTRUCTOR_TRAP_COL = "__leakage__touches_post_snapshot_21_90"
 
@@ -99,70 +105,8 @@ RENAME_MAP = {
 
 
 # ---------------------------------------------------------------------------
-# Pipeline steps
+# Version-specific pipeline steps
 # ---------------------------------------------------------------------------
-
-
-def derive_features(df: pd.DataFrame) -> pd.DataFrame:
-    """Derive binary and momentum features for the v7 column set."""
-    df = df.copy()
-    df["opportunity_created"] = df["opportunity_created"].astype(int)
-    df["demo_completed"] = (df["demo_page_views"] > 0).astype(int)
-    return df
-
-
-def softcap_expected_acv(
-    df: pd.DataFrame,
-    seed: int,
-    floor: float = ACV_FLOOR,
-    cap: float = ACV_CAP,
-) -> pd.DataFrame:
-    """Soft winsorize expected_acv to avoid hard-clipping ties at the cap.
-
-    Values below floor are clipped. Values above cap are pulled toward cap
-    with additive noise so they cluster near the cap rather than pile at it.
-    """
-    rng = RNGRoot(seed).numpy_child("softcap_acv")
-    df = df.copy()
-    acv = df["expected_acv"].copy()
-
-    # Floor: hard clip
-    acv = acv.clip(lower=floor)
-
-    # Cap: soft winsorize -- draw values in [cap - 5k, cap] for outliers
-    over_mask = acv > cap
-    n_over = int(over_mask.sum())
-    if n_over > 0:
-        acv.loc[over_mask] = cap - rng.uniform(0, 5000, size=n_over)
-
-    df["expected_acv"] = acv.round(0)
-    return df
-
-
-def assign_acquisition_wave(df: pd.DataFrame, seed: int) -> pd.DataFrame:
-    """Assign acquisition_wave (A, B, C) based on lead index position.
-
-    Waves A/B/C are roughly chronological: first third = A, middle = B,
-    last third = C. A small amount of noise is added at the boundaries.
-    """
-    rng = RNGRoot(seed).numpy_child("acquisition_wave")
-    df = df.copy()
-    n = len(df)
-    waves = np.empty(n, dtype=object)
-    third = n // 3
-
-    # Base assignment by position (stable across seeds)
-    waves[:third] = "A"
-    waves[third : 2 * third] = "B"
-    waves[2 * third :] = "C"
-
-    # Add ~5% boundary noise so it's not perfectly deterministic
-    noise_mask = rng.random(n) < 0.05
-    noise_vals = rng.choice(["A", "B", "C"], size=n)
-    waves[noise_mask] = noise_vals[noise_mask]
-
-    df["acquisition_wave"] = waves
-    return df
 
 
 def compute_post_snapshot_touches(
@@ -212,102 +156,11 @@ def rename_and_select(
         label_column: Source column for the binary label. Defaults to
             ``"converted_within_90_days"`` for backward compatibility.
     """
-    if label_column not in df.columns:
-        raise ValueError(
-            f"Label column {label_column!r} not found. Available: {sorted(df.columns)}"
-        )
-    if label_column == "converted_within_90_days":
-        rename_map = RENAME_MAP
-    else:
-        rename_map = {k: v for k, v in RENAME_MAP.items() if v != "converted"}
-        rename_map[label_column] = "converted"
-    df = df.rename(columns=rename_map)
-    df["converted"] = df["converted"].astype(int)
-    columns = FINAL_COLUMNS_INSTRUCTOR if instructor else FINAL_COLUMNS_STUDENT
-    missing = [c for c in columns if c not in df.columns]
-    if missing:
-        raise ValueError(
-            f"Missing required columns after renaming: {missing}. Available: {sorted(df.columns)}"
-        )
-    return df[columns]
-
-
-def subsample(
-    df: pd.DataFrame,
-    seed: int,
-    n: int = SUBSAMPLE_N,
-    target_rate: float = TARGET_RATE,
-) -> pd.DataFrame:
-    """Stratified subsample to n rows at target_rate conversion."""
-    rng = RNGRoot(seed).numpy_child("subsample")
-    positives = df[df["converted"] == 1]
-    negatives = df[df["converted"] == 0]
-    n_pos = int(n * target_rate)
-    n_neg = n - n_pos
-
-    if len(positives) < n_pos:
-        warnings.warn(
-            f"only {len(positives)} positives available, need {n_pos}",
-            stacklevel=2,
-        )
-        n_pos = len(positives)
-        n_neg = n - n_pos
-    if len(negatives) < n_neg:
-        warnings.warn(
-            f"only {len(negatives)} negatives available, need {n_neg}",
-            stacklevel=2,
-        )
-        n_neg = len(negatives)
-
-    pos_sample = positives.sample(n=n_pos, random_state=rng)
-    neg_sample = negatives.sample(n=n_neg, random_state=rng)
-    return (
-        pd.concat([pos_sample, neg_sample]).sample(frac=1, random_state=rng).reset_index(drop=True)
+    return _rename_and_select_generic(
+        df,
+        rename_map=RENAME_MAP,
+        final_columns=FINAL_COLUMNS_STUDENT,
+        instructor=instructor,
+        instructor_columns=FINAL_COLUMNS_INSTRUCTOR,
+        label_column=label_column,
     )
-
-
-def inject_missingness(df: pd.DataFrame, seed: int) -> pd.DataFrame:
-    """Apply structured missingness per the v7 contract.
-
-    Patterns:
-    1. Structural: days_since_last_touch is NaN when touch_count=0 (from snapshot)
-    2. MAR: web_sessions -- SDR outbound 15%, inbound marketing 2%, partner 5%
-    3. MAR: seniority -- partner referral 8%, others 1%
-    4. MCAR: expected_acv -- 2% uniform
-    5. Structural + MCAR: days_since_first_touch -- NaN when no touches + 2% MCAR
-    6. MCAR: days_since_last_touch -- additional 3% on top of structural
-    """
-    rng = RNGRoot(seed).numpy_child("missingness")
-    df = df.copy()
-    n = len(df)
-
-    # (1) Structural for days_since_last_touch is already NaN from snapshot builder
-    # Note: also structural for days_since_first_touch when no touches
-
-    # (2) MAR: web_sessions by lead_source
-    for source, rate in [
-        ("sdr_outbound", 0.15),
-        ("inbound_marketing", 0.02),
-        ("partner_referral", 0.05),
-    ]:
-        mask = (df["lead_source"] == source) & (rng.random(n) < rate)
-        df.loc[mask, "web_sessions"] = np.nan
-
-    # (3) MAR: seniority by lead_source
-    partner_mask = (df["lead_source"] == "partner_referral") & (rng.random(n) < 0.08)
-    other_mask = (df["lead_source"] != "partner_referral") & (rng.random(n) < 0.01)
-    df.loc[partner_mask | other_mask, "seniority"] = np.nan
-
-    # (4) MCAR: expected_acv 2%
-    acv_mcar = rng.random(n) < 0.02
-    df.loc[acv_mcar, "expected_acv"] = np.nan
-
-    # (5) MCAR: days_since_first_touch 2% on top of structural
-    dsft_mask = rng.random(n) < 0.02
-    df.loc[dsft_mask, "days_since_first_touch"] = np.nan
-
-    # (6) MCAR: days_since_last_touch 3% on top of structural
-    dslt_mask = rng.random(n) < 0.03
-    df.loc[dslt_mask, "days_since_last_touch"] = np.nan
-
-    return df
